# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: GeoGuess API
  version: 0.1.0
  summary: REST API contract for the GeoGuess-style game.
  description: >
    Versioned API contract for authentication, profiles, maps, games, guesses,
    rooms, parties, matchmaking, multiplayer matches, leaderboards, billing
    placeholders, and health checks. Exact location coordinates are never
    returned before a guess is locked or a multiplayer round is revealed.
    Unknown request properties are rejected. Missing and unauthorized party or
    match resources both return privacy-safe not_found where revealing existence
    would leak information.
servers:
  - url: http://localhost/api/v1
    description: Local development through Nginx
  - url: http://localhost:8080/api/v1
    description: Local Go API direct
  - url: https://api.example.com/api/v1
    description: Production placeholder
security: []
...
paths:
...
  /users/{userId}/games:
    get:
      tags: [Profile]
      summary: Get user game history
      operationId: getUserGameHistory
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Cursor-paginated, public-safe list of the user's games.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGameHistoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: User is missing or unavailable. Response does not distinguish private account states.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /games:
    post:
      tags: [Games]
      summary: Create a game
      description: Creates a Phase 04 solo game for a registered or guest session. The backend selects all round locations, stores durable game state, and rejects maps without enough unique active locations.
      operationId: createGame
      security:
        - accessCookie: []
          csrfHeader: []
        - guestSession: []
          csrfHeader: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGameRequest'
      responses:
        '201':
          description: Game created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /games/{gameId}:
    get:
      tags: [Games]
      summary: Get game state
      operationId: getGame
      security:
        - accessCookie: []
        - guestSession: []
      parameters:
        - $ref: '#/components/parameters/GameId'
      responses:
        '200':
          description: Game state visible to the participant.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /games/{gameId}/start:
    post:
      tags: [Games]
      summary: Start a game
      description: Starts a pending solo game, sets server-owned timestamps, and activates the first round.
      operationId: startGame
      security:
        - accessCookie: []
          csrfHeader: []
        - guestSession: []
          csrfHeader: []
      parameters:
        - $ref: '#/components/parameters/GameId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Game started.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /games/{gameId}/rounds/current:
    get:
      tags: [Games]
      summary: Get current round
      description: Returns current round data without true latitude, true longitude, location id, provider refs, or any hidden answer metadata.
      operationId: getCurrentRound
      security:
        - accessCookie: []
        - guestSession: []
      parameters:
        - $ref: '#/components/parameters/GameId'
      responses:
        '200':
          description: Current round.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentRoundResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /games/{gameId}/rounds/{roundId}/guesses:
    post:
      tags: [Games]
      summary: Submit a guess
      description: >
        Submits one server-scored guess for the current round. Solo/daily responses
        reveal the answer immediately. Multiplayer responses keep actual_location
        null until the shared round closes (all eligible submissions or ranked
        deadline). Repeating the same Idempotency-Key with the same coordinates
        returns the original result; reusing it with different coordinates returns
        conflict (idempotency_conflict).
      operationId: submitGuess
      security:
        - accessCookie: []
          csrfHeader: []
        - guestSession: []
          csrfHeader: []
      parameters:
        - $ref: '#/components/parameters/GameId'
        - $ref: '#/components/parameters/RoundId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitGuessRequest'
            examples:
              cairo:
                value:
                  latitude: 30.0444
                  longitude: 31.2357
      responses:
        '200':
          description: Guess accepted or idempotently replayed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuessResultResponse'
              examples:
                soloRevealed:
                  summary: Solo/daily immediate reveal
                  value:
                    guess:
                      id: 00000000-0000-4000-8000-0000000000a1
                      latitude: 30.0444
                      longitude: 31.2357
                      distance_meters: 12345
                      accuracy_score: 4020
                      speed_bonus: 0
                      score: 4020
                      submitted_at: "2026-07-18T10:02:25Z"
                      timed_out: false
                    round_completed: true
                    game_completed: false
                    actual_location:
                      latitude: 30.05
                      longitude: 31.24
                      country_code: EG
                    max_score: 5000
                    score_percent: 80
                    max_accuracy_score: 5000
                    max_speed_bonus: 250
                    outcome: close
                multiplayerDelayedReveal:
                  summary: Multiplayer delayed reveal (round still open)
                  value:
                    guess:
                      id: 00000000-0000-4000-8000-0000000000a2
                      latitude: 48.8566
                      longitude: 2.3522
                      distance_meters: 5432
                      accuracy_score: 4120
                      speed_bonus: 0
                      score: 4120
                      submitted_at: "2026-07-18T10:02:25Z"
                      timed_out: false
                    round_completed: false
                    game_completed: false
                    submitted_count: 3
                    eligible_count: 4
                    actual_location: null
                    max_score: 5000
                    score_percent: 82
                    max_accuracy_score: 5000
                    max_speed_bonus: 250
                    outcome: submitted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /games/{gameId}/rounds/{roundId}/timeout:
    post:
      tags: [Games]
      summary: Advance an expired daily round
      description: Records a zero-score timeout after the server deadline, reveals the correct location, and activates the next round or completes the game.
      operationId: expireDailyRound
      security:
        - accessCookie: []
          csrfHeader: []
      parameters:
        - $ref: '#/components/parameters/GameId'
        - $ref: '#/components/parameters/RoundId'
      responses:
        '200':
          description: Expired round advanced.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuessResultResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /games/{gameId}/results:
    get:
      tags: [Games]
      summary: Get game results
      description: Returns durable final results for the owning player after a solo game is completed.
      operationId: getGameResults
      security:
        - accessCookie: []
        - guestSession: []
      parameters:
        - $ref: '#/components/parameters/GameId'
      responses:
        '200':
          description: Game results.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResultsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /rooms/{roomCode}:
    get:
      tags: [Rooms]
      summary: Get room state
      operationId: getRoom
      security:
        - accessCookie: []
        - guestSession: []
      parameters:
        - $ref: '#/components/parameters/RoomCode'
      responses:
        '200':
          description: Room state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    accessCookie:
      type: apiKey
      in: cookie
      name: access_token
      description: Short-lived JWT access token in an HTTP-only secure cookie.
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: Rotated refresh token in an HTTP-only secure cookie.
    guestSession:
      type: apiKey
      in: cookie
      name: guest_session
      description: Signed guest session cookie for anonymous gameplay.
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required for unsafe cookie-authenticated requests.
    paymentWebhookSignature:
      type: apiKey
      in: header
      name: X-Payment-Signature
      description: Provider-specific payment webhook signature.
...
  parameters:
...
    GameId:
      name: gameId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RoundId:
      name: roundId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    RoomCode:
      name: roomCode
      in: path
      required: true
      schema:
        type: string
        pattern: '^[A-Z0-9]{6,10}$'
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
        minLength: 1
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Stable retry key for replay-safe writes. Identical requests replay the original successful result; conflicting reuse returns 409.
      schema:
        type: string
        minLength: 16
        maxLength: 128
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated but not allowed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Resource conflict or duplicate command.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity:
      description: Request is syntactically valid but violates domain state rules.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          examples: [validation_failed]
        message:
          type: string
          examples: [The request body is invalid.]
        request_id:
          type: string
          examples: [req_01J0H5ZK9Y6B4D7T2M9F3X1A8C]
        fields:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      required: [name, code, message]
      properties:
        name:
          type: string
          examples: [email]
        code:
          type: string
          examples: [invalid_email]
        message:
          type: string
          examples: [Email must be valid.]
    PageInfo:
      type: object
      required: [limit]
      properties:
        limit:
          type: integer
          minimum: 1
          maximum: 100
        next_cursor:
          type: [string, 'null']
...
    UserGameHistoryResponse:
      type: object
      required: [games, page]
      properties:
        games:
          type: array
          items:
            $ref: '#/components/schemas/UserGameHistoryItem'
        page:
          $ref: '#/components/schemas/PageInfo'
    UserGameHistoryItem:
      type: object
      required: [id, map_id, mode, status, round_count, total_score, created_at]
      properties:
        id:
          type: string
          format: uuid
        map_id:
          type: string
          format: uuid
        mode:
          type: string
        status:
          type: string
        round_count:
          type: integer
          minimum: 1
        current_round_number:
          type: [integer, 'null']
          minimum: 1
        total_score:
          type: integer
          minimum: 0
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
...
    CreateGameRequest:
      type: object
      required: [mode, map_id, round_count]
      additionalProperties: false
      description: Phase 04 accepts solo games only, even though the shared GameMode enum includes future modes.
      properties:
        mode:
          $ref: '#/components/schemas/GameMode'
        map_id:
          type: string
          format: uuid
        round_count:
          type: integer
          minimum: 1
          maximum: 10
          default: 5
        timer_seconds:
          type: [integer, 'null']
          minimum: 10
          maximum: 600
    GameResponse:
      type: object
      required: [game]
      properties:
        game:
          $ref: '#/components/schemas/Game'
    Game:
      type: object
      required:
        - id
        - mode
        - status
        - round_count
        - scoring_version
        - current_round_number
        - total_score
      properties:
        id:
          type: string
          format: uuid
        mode:
          $ref: '#/components/schemas/GameMode'
        status:
          $ref: '#/components/schemas/GameStatus'
        map_id:
          type: string
          format: uuid
        round_count:
          type: integer
        timer_seconds:
          type: [integer, 'null']
        scoring_version:
          type: integer
        current_round_number:
          type: [integer, 'null']
        total_score:
          type: integer
          minimum: 0
        started_at:
          type: [string, 'null']
          format: date-time
        completed_at:
          type: [string, 'null']
          format: date-time
    GameMode:
      type: string
      enum: [solo, private_room, quick_play, daily, ranked]
    GameStatus:
      type: string
      enum: [pending, active, completed, abandoned, cancelled]
    CurrentRoundResponse:
      type: object
      required: [round]
      properties:
        round:
          $ref: '#/components/schemas/Round'
    Round:
      type: object
      required: [id, round_number, status, media]
      properties:
        id:
          type: string
          format: uuid
        round_number:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [pending, active, completed, cancelled]
        starts_at:
          type: [string, 'null']
          format: date-time
        ends_at:
          type: [string, 'null']
          format: date-time
        media:
          $ref: '#/components/schemas/RoundMedia'
    SubmitGuessRequest:
      type: object
      required: [latitude, longitude]
      additionalProperties: false
      properties:
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
    GuessResultResponse:
      type: object
      required: [guess, max_score, score_percent, outcome]
      description: >
        Solo/daily always populate actual_location after accept. Multiplayer keeps
        actual_location null while round_completed is false. Compatibility aliases
        max_score and score_percent remain until solo clients migrate to
        max_accuracy_score / accuracy fields.
      properties:
        guess:
          $ref: '#/components/schemas/GuessResult'
        actual_location:
          anyOf:
            - $ref: '#/components/schemas/RevealedLocation'
            - type: 'null'
          description: Null for multiplayer until the shared round closes.
        max_score:
          type: integer
          const: 5000
          description: Legacy alias for max accuracy score (solo/daily compatibility).
        score_percent:
          type: integer
          minimum: 0
          maximum: 100
        max_accuracy_score:
          type: integer
          minimum: 0
          maximum: 5000
          description: Maximum accuracy points per round (typically 5000).
        max_speed_bonus:
          type: integer
          minimum: 0
          maximum: 250
          description: Maximum speed bonus per round (0 for casual; up to 250 ranked).
        outcome:
          type: string
          description: >
            Solo outcomes include perfect/close/miss/timed_out. Multiplayer may
            return submitted while the shared round remains open.
          enum: [perfect, close, miss, timed_out, submitted]
        round_completed:
          type: boolean
          description: True when the accepted guess closes the shared or solo round.
        game_completed:
          type: boolean
        submitted_count:
          type: integer
          minimum: 0
          description: Multiplayer aggregate of locked guesses for the current round.
        eligible_count:
          type: integer
          minimum: 0
          description: Multiplayer count of eligible participants for the current round.
        next_round_number:
          type: integer
          minimum: 1
    GuessResult:
      type: object
      required: [id, latitude, longitude, distance_meters, score, submitted_at, timed_out]
      properties:
        id:
          type: string
          format: uuid
        latitude:
          type: number
        longitude:
          type: number
        distance_meters:
          type: integer
          minimum: 0
        accuracy_score:
          type: integer
          minimum: 0
          maximum: 5000
          description: Distance-based accuracy points before speed bonus.
        speed_bonus:
          type: integer
          minimum: 0
          maximum: 250
          description: Ranked remaining-time bonus; zero for casual.
        score:
          type: integer
          minimum: 0
          maximum: 5250
          description: Accuracy plus speed bonus for this guess.
        submitted_at:
          type: string
          format: date-time
        timed_out:
          type: boolean
    RevealedLocation:
      type: object
      required: [latitude, longitude, country_code]
      properties:
        latitude:
          type: number
        longitude:
          type: number
        country_code:
          type: string
          minLength: 2
          maxLength: 2
        region:
          type: [string, 'null']
        locality:
          type: [string, 'null']
    GameResultsResponse:
      type: object
      required: [game, players, rounds]
      properties:
        game:
          $ref: '#/components/schemas/Game'
        players:
          type: array
          items:
            $ref: '#/components/schemas/GamePlayer'
        rounds:
          type: array
          items:
            $ref: '#/components/schemas/RoundResult'
    GamePlayer:
      type: object
      required: [id, display_name, role, status, total_score]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: [string, 'null']
          format: uuid
        display_name:
          type: string
        role:
          type: string
          enum: [host, player, spectator]
        status:
          type: string
          enum: [active, disconnected, left, kicked]
        total_score:
          type: integer
          minimum: 0
    RoundResult:
      type: object
      required: [round_id, round_number, actual_location, guesses]
      properties:
        round_id:
          type: string
          format: uuid
        round_number:
          type: integer
        actual_location:
          $ref: '#/components/schemas/RevealedLocation'
        guesses:
          type: array
          items:
            $ref: '#/components/schemas/GuessResult'
...
    RoomResponse:
      type: object
      required: [room]
      properties:
        room:
          $ref: '#/components/schemas/Room'
...